requirements.txt
certifi==2019.3.9
Django==2.2
djangorestframework==3.10.1
gunicorn==19.9.0
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
urllib3==1.24.1
